files/bone/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
httpx>=0.27
PyYAML>=6.0
